Addressed feedbacks
