+++ internal/pkg/cli/completion.go
@@ -11,7 +11,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/aws/PRIVATE-amazon-ecs-archer/cmd/archer/template"
-	"github.com/aws/PRIVATE-amazon-ecs-archer/internal/pkg/cli/groups"
+	"github.com/aws/PRIVATE-amazon-ecs-archer/internal/pkg/cli/group"
 )
 
 type shellCompleter interface {
@@ -80,7 +80,7 @@ The code must be evaluated to provide interactive completion of commands.`,
 	}
 	cmd.SetUsageTemplate(template.Usage)
 	cmd.Annotations = map[string]string{
-		"group": groups.Settings,
+		"group": group.Settings,
 	}
 	return cmd
 }
