+++ internal/pkg/cli/group/group.go
@@ -1,11 +1,13 @@
+// +build !windows
+
 // Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
 // Package groups contains the names of command groups
-package groups
+package group
 
 const (
-	GettingStarted = "Getting Started"
-	Develop = "Develop"
+	GettingStarted = "Getting Started ✨"
+	Develop = "Develop 🔧"
 	Settings = "Settings"
 )
