+++ internal/pkg/cli/project.go
@@ -7,7 +7,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/aws/PRIVATE-amazon-ecs-archer/cmd/archer/template"
-	"github.com/aws/PRIVATE-amazon-ecs-archer/internal/pkg/cli/groups"
+	"github.com/aws/PRIVATE-amazon-ecs-archer/internal/pkg/cli/group"
 )
 
 // BuildProjCmd builds the top level project command and related subcommands.
@@ -22,7 +22,7 @@ A Project represents all of your deployment environments.`,
 	cmd.AddCommand(BuildProjectListCommand())
 	cmd.SetUsageTemplate(template.Usage)
 	cmd.Annotations = map[string]string{
-		"group": groups.Develop,
+		"group": group.Develop,
 	}
 	return cmd
 }
